fix(ci): keep zone runner branch linear and lint-clean
